+++ cookbooks/dokuwiki/attributes/default.rb
@@ -1,0 +1,6 @@
+default['dokuwiki']['db_user'] = 'hello_app'
+default['dokuwiki']['db_name'] = 'hello_app'
+default['dokuwiki']['server_name'] = 'wiki.headwire.com'
+default['dokuwiki']['docroot'] = '/srv/wiki/current'
+default['dokuwiki']['config_dir'] = '/srv/wiki/shared/config'
+