+++ cookbooks/dokuwiki/metadata.rb
@@ -1,0 +1,10 @@
+name             'dokuwiki'
+maintainer       'Kyle Watson'
+maintainer_email 'kw@headwire.com'
+license          'All rights reserved'
+description      'Installs/Configures dokuwiki'
+long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+version          '0.1.0'
+
+depends          "apache2"
+depends          "php"