+++ cookbooks/dokuwiki/recipes/default.rb
@@ -1,8 +1,10 @@
 #
-# Cookbook Name:: headwire
+# Cookbook Name:: dokuwiki
 # Recipe:: default
 #
 # Copyright 2014, YOUR_COMPANY_NAME
 #
 # All rights reserved - Do Not Redistribute
 #
+
+include_recipe "dokuwiki::webserver"