+++ environments/dev.rb
@@ -1,0 +1,6 @@
+override_attributes ({
+  "dokuwiki" => {
+    "server_name" => "dokuwiki.dev",
+    "docroot" => "/home/vagrant/apps/dokuwiki"
+  }
+})